open source 09df54c0cc99354a60bbc0303e3e8ea33a96bef0 (#2725)

+++ scripts/build_wheel.py
@@ -86,7 +86,8 @@ def main(*,
          python_bindings: bool = True,
          benchmarks: bool = False,
          micro_benchmarks: bool = False,
-         nvtx: bool = False):
+         nvtx: bool = False,
+         skip_stubs: bool = False):
 
     if clean:
         clean_wheel = True
@@ -256,8 +257,12 @@ def main(*,
             lib_dir / "libtensorrt_llm_ucx_wrapper.so")
         copy(
             build_dir /
-            "tensorrt_llm/kernels/decoderMaskedMultiheadAttention/libdecoder_attention.so",
-            lib_dir / "libdecoder_attention.so")
+            "tensorrt_llm/kernels/decoderMaskedMultiheadAttention/libdecoder_attention_0.so",
+            lib_dir / "libdecoder_attention_0.so")
+        copy(
+            build_dir /
+            "tensorrt_llm/kernels/decoderMaskedMultiheadAttention/libdecoder_attention_1.so",
+            lib_dir / "libdecoder_attention_1.so")
 
     bin_dir = pkg_dir / "bin"
     if bin_dir.exists():
@@ -283,47 +288,48 @@ def main(*,
             return pybind_lib[0]
 
         copy(get_pybind_lib(), pkg_dir)
+        if not skip_stubs:
+            with working_directory(project_dir):
+                build_run(
+                    f"\"{sys.executable}\" -m pip install pybind11-stubgen")
+            with working_directory(pkg_dir):
+                if on_windows:
+                    stubgen = "stubgen.py"
+                    stubgen_contents = """
+                    # Loading torch, trt before bindings is required to avoid import errors on windows.
+                    # isort: off
+                    import torch
+                    import tensorrt as trt
+                    # isort: on
+                    import os
+                    import platform
 
-        with working_directory(project_dir):
-            build_run(f"\"{sys.executable}\" -m pip install pybind11-stubgen")
-        with working_directory(pkg_dir):
-            if on_windows:
-                stubgen = "stubgen.py"
-                stubgen_contents = """
-                # Loading torch, trt before bindings is required to avoid import errors on windows.
-                # isort: off
-                import torch
-                import tensorrt as trt
-                # isort: on
-                import os
-                import platform
+                    from pybind11_stubgen import main
 
-                from pybind11_stubgen import main
+                    if __name__ == "__main__":
+                        # Load dlls from `libs` directory before launching bindings.
+                        if platform.system() == "Windows":
+                            os.add_dll_directory(r\"{lib_dir}\")
+                        main()
+                    """.format(lib_dir=lib_dir)
+                    (pkg_dir / stubgen).write_text(dedent(stubgen_contents))
+                    build_run(f"\"{sys.executable}\" {stubgen} -o . bindings")
+                    (pkg_dir / stubgen).unlink()
+                else:
+                    env_ld = os.environ.copy()
 
-                if __name__ == "__main__":
-                    # Load dlls from `libs` directory before launching bindings.
-                    if platform.system() == "Windows":
-                        os.add_dll_directory(r\"{lib_dir}\")
-                    main()
-                """.format(lib_dir=lib_dir)
-                (pkg_dir / stubgen).write_text(dedent(stubgen_contents))
-                build_run(f"\"{sys.executable}\" {stubgen} -o . bindings")
-                (pkg_dir / stubgen).unlink()
-            else:
-                env_ld = os.environ.copy()
-
-                new_library_path = "/usr/local/cuda/compat/lib.real"
-                if 'LD_LIBRARY_PATH' in env_ld:
-                    new_library_path += f":{env_ld['LD_LIBRARY_PATH']}"
-                env_ld["LD_LIBRARY_PATH"] = new_library_path
-                try:
-                    build_run(
-                        f"\"{sys.executable}\" -m pybind11_stubgen -o . bindings --exit-code",
-                        env=env_ld)
-                except CalledProcessError as ex:
-                    print(f"Failed to build pybind11 stubgen: {ex}",
-                          file=sys.stderr)
-                    exit(1)
+                    new_library_path = "/usr/local/cuda/compat/lib.real"
+                    if 'LD_LIBRARY_PATH' in env_ld:
+                        new_library_path += f":{env_ld['LD_LIBRARY_PATH']}"
+                    env_ld["LD_LIBRARY_PATH"] = new_library_path
+                    try:
+                        build_run(
+                            f"\"{sys.executable}\" -m pybind11_stubgen -o . bindings --exit-code",
+                            env=env_ld)
+                    except CalledProcessError as ex:
+                        print(f"Failed to build pybind11 stubgen: {ex}",
+                              file=sys.stderr)
+                        exit(1)
 
     if not skip_building_wheel:
         if dist_dir is None:
@@ -429,6 +435,9 @@ def add_arguments(parser: ArgumentParser):
     parser.add_argument("--nvtx",
                         action="store_true",
                         help="Enable NVTX features.")
+    parser.add_argument("--skip-stubs",
+                        action="store_true",
+                        help="Skip building python stubs")
 
 
 if __name__ == "__main__":
